profile info and quest GUI
Added basic GUI for profile info and quest. A quest selection GUI, a
profile selection routine and panel for quick profile and quest info.
Quests and profiles are not yet implemented.

Added breakpoint support in the GUI. Breakpoints are at the moment not
activated.
