+++ src/at/jku/ssw/cmm/gui/event/debug/PanelToolBarListener.java
@@ -1,20 +1,20 @@
-package at.jku.ssw.cmm.gui.event.panel;
+package at.jku.ssw.cmm.gui.event.debug;
 
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
 
-import at.jku.ssw.cmm.gui.GUIrightPanel;
+import at.jku.ssw.cmm.gui.GUIdebugPanel;
 import at.jku.ssw.cmm.gui.mod.GUImainMod;
 
 public class PanelToolBarListener {
 
-	public PanelToolBarListener( GUImainMod modifier, GUIrightPanel master ){
+	public PanelToolBarListener( GUImainMod modifier, GUIdebugPanel master ){
 		this.modifier = modifier;
 		this.master = master;
 	}
 	
 	private final GUImainMod modifier;
-	private final GUIrightPanel master;
+	private final GUIdebugPanel master;
 	
 	public MouseListener breakPointButtonHandler = new MouseListener() {
